@@ -2,16 +2,14 @@
 import * as React from "react";
 
 import { ScrollArea } from "@/components/ui/scroll-area";
-import { Separator } from "@/components/ui/separator";
 import SidebarLink from "@/components/ui/sidebar-link";
-import { usePathname } from "next/navigation";
 
 export function Sidebar() {
   return (
     <div className="fixed h-full">
-      <ScrollArea className=" w-60 h-full pl-4">
+      <ScrollArea className=" h-full w-60 pl-4">
         <div className="p-4">
-          <h4 className="font-semibold text-sm">Hooks</h4>
+          <h4 className="text-sm font-semibold">Hooks</h4>
           <ul className="mt-2">
             <li className="my-4">
               <SidebarLink href="/documentation/useDelayedLoading">
